chore: bind Hugo server to 0.0.0.0
By default Hugo would bind to `localhost`/`127.0.0.1`, when developing
the website inside the container and exposing the `1313` port this
becomes a problem as a port bound to `localhost` can't be exposed.

This adds the `--bind 0.0.0.0` so that Hugo's server binds to all
interfaces/ip addresses available. This allows the port to be exposed
when running within a container.