chore: update docs
This brings over updates made separately in the Syndesis Developer
Handbook in the main Syndesis repository, updates the documents to
replace outdated information and reformats the docs a bit.

I think the docs still are not 100% up to date, at least this would help
in removing the need for parallel documentation in the Syndesis
Developer Handbook and the confusion about where to edit the docs and
where to point users/contributors for documentation.