refactor: flexible layout and TOC styling
This makes sure that the TOC takes up as much space as it needs to to
prevent word wrap (as much as possible), making the TOC a bit
unreadable by tweaking the flex layout to have fixed 214px sidenav,
growable TOC and main content taking up remaining space.

For TOC styling this removes the gray background, use of italic for
headings; and increases the line height to make the menu more legible.
Also adds a gradient as a separator from the content.