chore: quick and dirty Asciidoc rendering
To add the Syndesis chrome (header/footer) this adds a Asciidoctor.js
Docinfo processor with hardcoded HTML for header and footer. For styling
making sure that the TOC is on the left hand side and it doesn't overlap
the content bits of hacky CSS were added. Manual link was removed from
the top navigation and the content was added to `docs/index.html`.
Asciidoctor.js now uses the `syndesis.css` file instead of embedding the
default CSS theme.